+++ tests/test_gossip.py
@@ -1012,6 +1012,11 @@ def test_node_reannounce(node_factory, bitcoind):
     wait_for(lambda: 'alias' in only_one(l2.rpc.listnodes(l1.info['id'])['nodes']))
     assert only_one(l2.rpc.listnodes(l1.info['id'])['nodes'])['alias'].startswith('JUNIORBEAM')
 
+    lfeatures = '28a2'
+
+    # Make sure it gets features correct.
+    assert only_one(l2.rpc.listnodes(l1.info['id'])['nodes'])['globalfeatures'] == lfeatures
+
     l1.stop()
     l1.daemon.opts['alias'] = 'SENIORBEAM'
     # It won't update within 5 seconds, so sleep.
@@ -1429,7 +1434,7 @@ def test_gossip_store_compact_on_load(node_factory, bitcoind):
     l2.restart()
 
     wait_for(lambda: l2.daemon.is_in_log(r'gossip_store_compact_offline: [5-8] deleted, 9 copied'))
-    wait_for(lambda: l2.daemon.is_in_log(r'gossip_store: Read 1/4/2/0 cannounce/cupdate/nannounce/cdelete from store \(0 deleted\) in 1446 bytes'))
+    wait_for(lambda: l2.daemon.is_in_log(r'gossip_store: Read 1/4/2/0 cannounce/cupdate/nannounce/cdelete from store \(0 deleted\) in 1450 bytes'))
 
 
 def test_gossip_announce_invalid_block(node_factory, bitcoind):